+++ tests/testthat/test_sampleParsSimpar.R
@@ -12,7 +12,7 @@ test_that("basic",{
 
         fileRef <- "testReference/sampleParsSimpar_01.rds"
 
-        pars <- sampleParsSimpar(file.mod,nsim=10,format="wide",as.fun="data.table",seed.R=23)
+        pars <- expect_warning(sampleParsSimpar(file.mod,nsim=10,format="wide",as.fun="data.table",seed.R=23))
 
         expect_equal_to_reference(pars,fileRef)
 
@@ -53,7 +53,9 @@ test_that("mvrnorm through samplePars()",{
     fileRef <- "testReference/samplePars_03.rds"
     
     pars.mvrnorm <- samplePars(file.mod=file.mod,nsims=10,method="mvrnorm",format="ext",as.fun="data.table",seed.R=23)
-    
+
+    cols.round <- c("value","value.est")
+    pars.mvrnorm[,(cols.round):=lapply(.SD,round,3),.SDcols=cols.round]
     expect_equal_to_reference(pars.mvrnorm,fileRef)
     
 })