+++ NEWS.md
@@ -1,4 +1,8 @@
 # NMsim 0.2.2
+## New features
+* `sampleCovs()` is a new function to conveniently sample
+  subject-level covariates from an existing data set.
+
 ## Bugfixes
 * There was a bug introduced in version 0.2.1 leading `NMsim()` to
 throw an error saving some data sets. The issue is fixed.